[ADD] web_tour_recorder: add a tour recorder
A new module 'web_tour_recorder' that allow to create tours
on the fly. Like a macro recorder.
You can export the created tour to JSON to re-import it on
another pc or database for example.
You can also export it to JS to have a tour file (almost) ready
to be put in the tour folder of your development.

The recorder can be started through the debug menu.
The custom tours are stored in the localStorage.

closes odoo/odoo#153791

Task-id: 3927002
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>