[FIX] web_tour: correctly log failed step
When a tour fails, the tour system is supposed to log the failing
step (with the 3 previous/next steps). Before this commit, this
didn't work, and the first 3 steps were always logged, no matter
which step failed.

This didn't work because to determine the failing step, we try to
find the step from the list of steps, with reference matching
(steps are objects).

However, since [1], steps are obtained from a getter which calls
the steps function of the tour, so they always get a new version
of the steps.

This commit fixes the issue by memoizing the steps, such that the
function is called only once, and we keep the same references to
the step objects.

Note that this could be reworked in master to make it more robust
(e.g. finding steps based on ids).

[1] https://github.com/odoo/odoo/commit/6dec75c23446a0dcb24467b0d9d3f03d35c95936

closes odoo/odoo#145709

X-original-commit: 2377a33b8b182cc0160ae4bbe1cdbbe9a7d58393
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>