[IMP] web,*: custom success signal for js tests
*im_livechat,point_of_sale,web_tour,tests

This commit introduces specific success signals for each type of
js tests (tours, qunit, hoot, clickbot). Before, they all waited
for the string "test successful" to be logged. This already led to
a very problematic issue, which was that the qunit suite was no
longer run entirely (see [1] for details).

With those custom signals, we reduce the risk of interferences
between those types of tests.

This commit also makes the success_signal argument of browser_js
accept strings, which is more convenient when we want an exact
string matching. By default, it still expect "test successful",
but whereas before it waited for a string containing "test
successful", it now waits for this exact log.

[1] 4e7a674348c5e3460f0593796d75a3887e62a3f0

task 3665259

Part-of: odoo/odoo#153378