[FIX] web: HOOT - prevent setTimeout overrides
This commit makes the global and the browser module's `setTimeout`
methods and other related methods read-only.

This is because HOOT is handling timing functions in such a way that
overriding them risks breaking its internal logic and could cause
concurrency issues in the test runner.

The motivation is that there are already helpers that should cover all
use cases where timers need to be triggered.

This commit also allows the `runAllTimers` to accept registering other
timing functions.

Part-of: odoo/odoo#155073