[FIX] web: add missing await in mock server
There was a missing `await` statement in front of a `request.json()`
call which is supposed to be asynchronous.

It didn't throw an error since the mock response object's methods are
rendered synchronous to increase tests performances, but it is wrong
regardless and messes up type inference.

Part-of: odoo/odoo#156869