[IMP] web: tests - centralize 'get_definitions'
This commit ensures that a single call to the server `get_definitions`
is performed for all models that will be needed in the current test run
at the start of the run.

It also allows models fetched from the server to ignore inheritance from
models that have not been fetched.

Part-of: odoo/odoo#159826