[FIX] web: runbot should not timeout when a test fails
Before this commit, when a test failed, runbot would wait until a
message does not contain "[HOOT]" and timout.

Now, it will wait until a message does not contain "[HOOT]" or a message
that contains "[HOOT] test failed (see above for details)". This means
that the runbot will not time out anymore.

closes odoo/odoo#160891

X-original-commit: eed70af6a8907d1a3c93ddf5a7f0f80c0d2f8305
Signed-off-by: Julien Mougenot (jum) <jum@odoo.com>
Signed-off-by: Jorge Pinna Puissant (jpp) <jpp@odoo.com>