[FIX] web_tour: observe shadow roots when resuming tour
Prior to this commit, the tour service would fail to observe changes
in the tour's shadow DOM after the tour was resumed.

A mutation observer waits for the shadow host to appear in the DOM,
then registers it with the macro engine observer in order for the
macro engine to take shadow DOM modifications into account.

This commit ensures that the same logic is applied when resuming a
tour.

X-original-commit: edf4e8086a6210e5a57264d72e9341ba8691d320
Part-of: odoo/odoo#161511