[IMP] web, base: Add embedded actions
This update introduces the implementation of embedded actions, enabling
the addition of embedded actions in the control panel for seamless
toggling among various actions.

Functional specifications:

This task aims to centralize and access models related to the current
parent model.

We introduce the concept of embedded action. If an action has embedded
actions, then a button with an 'sliders' icon is displayed in the control
panel.

When the user clicks the button, the embedded actions appears.

The implementation consists as various actions chosen as visible
by the user. By default, the embedded actions only displays the parent
action (and will always display it).

There is a dropdown button to the right of the topbar that displays
all the actions available to the user. These actions depend on the model
, but it's also possible that there is a domain on these actions making
them invisible. For example, a project without the 'allow timesheets'
option will not display the 'Timesheets' action.

The default actions cannot be deleted by the user; they can only
decide whether to display them or not. The visibility of actions is
different for each model but also for each model ID, meaning, for
example, that the actions visible in one project will not be the same
for another.

Each user can also save the current view and thus create a new embedded
action. They can decide to share this embedded action so that it is
visible to everyone, or not. The view type and filters are saved when
creating a new embedded action. Those saved filters are saved as favorites
filters that cannot be deleted while the embedded action exists.

If the user has the rights, he can also drag and drop the actions to
reorder them. This order will be specific to the user and to the model
ID in which the reordering was made.

On Mobile, only a dropdown is displayed with the current action. When
the dropdown is unfolded, all available actions are displayed. It is
not possible to drag and drop. Deleting is always available, as well as
the creation of new actions.

The breadcrumbs should never push new items. When navigating between
embedded actions, the breadcrumbs should be replaced.

In order to create a new default embedded action for a model, the
developer has to create an xml record of that structure:

<record id="<id>" model="ir.embedded.actions">
    <field name="parent_res_model"><parent_res_model></field>
    <field name="parent_action_id" ref="<xml_id_parent_action>"/>
    <!--
        for the triggered action, the developer can either choose
        between a xml action id or a python action
    -->
    <field name="action_id" ref=<triggered_action_id>/>
    <field name="python_method"><python_method></field>
    <!-- name is optional if an action_id is declared -->
    <field name="name"><name_of_the_embedded_action></field>
    <!-- optional -->
    <field name="sequence"><sequence_order_integer></field>
    <field name="domain"><domain_to_filter_embedded_actions></field>
    <field name="context"><additional_context_keys></field>
    <field name="default_view_mode"><default_view></field>
    <field name="filter_ids"><Filters></field>
    <field name="groups_ids"><Groups></field>
</record>

taskid:3750721