[FIX] web: uniformise action load errors when the action is not found
When searching an action, ones can use the XML ID; the path or the ID of
the action. If the XML ID or the path is used, and the action is not
found, a missing error is raised. Contrariwise, if the ID of the action
is used, and the action is not found, an empty record is returned.

This difference of behaviour, when treating a not found action, could
cause some errors. For instance, the JS code expects to receive always
an error to show the corresponding notification, when receiving an empty
record instead, this empty action is used to continue to execute the
code and other errors are raised.

Now, if the record is not found, we always raise an error.

closes odoo/odoo#165599

X-original-commit: bde1c50bb41b7a2cd302bdd995eace7428d22026
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>
Signed-off-by: Jorge Pinna Puissant (jpp) <jpp@odoo.com>