[FIX] web: reload webclient when activating developer mode
- Settings;
- Activate the developer mode (with assets);

Before this commit, the query string `?debug=1` is correctly added to
the url, but the developer mode is not active because the webclient
didn't reload.

This occurs because, since [1], internal links that point to a different
place on the same Odoo instance will simply load the URL without causing
a full page reload.

Now, the webclient is correctly reloaded when activating or deactivating
the developer mode.

[1] https://github.com/odoo/odoo/commit/3ebd4d764dd241c292594664acbc1f366bf69b4f

opw-3945721

closes odoo/odoo#166774

X-original-commit: 5e714e15551c7803fdcba0570cac2da8d0ccd1b6
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>
Signed-off-by: Jorge Pinna Puissant (jpp) <jpp@odoo.com>