[FIX] web_tour: isCheck: do not allow invisible elements
During previous PRs, a typo was introduced in the tour_compiler (in
commit 911396c7e49aa69da940a08154e476a20d9ddbc9)
regarding the behavior of the isCheck attribute for a step of a tour.
Before this commit, isCheck: true accepted that the trigger was invisible.
After this commit, isCheck: true no longer accepts that the trigger is
invisible.
Attention !!! isCheck: true is not equal to run() {}.
- isCheck: true => Accepts that the trigger is disabled.
- run() {} => DOES NOT accept that the trigger is disabled.

closes odoo/odoo#166527

Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>