[REF] web_tour: remove shadow_dom attribute
In order to significantly simplify the turn step API, we are removing
the "shadow_dom" attribute. Indeed, it is possible to select the trigger
with a pseudo selector (from hoot) in a shadow_dom by adding ":shadow"
to the trigger. From then on, this functionality no longer needs to be
supported by the tour_compiler.

task~3974087

closes odoo/odoo#168640

Signed-off-by: Julien Mougenot (jum) <jum@odoo.com>