[REF] web_tour: remove compileToMacro method
In this commit, we remove compileToMacro method from TourStepAutomatic
class. The compileToMacro method complicates the understanding of how
tours work. It is more obvious and above all simpler to directly return
an object corresponding to the macro input (in TourAutomatic) rather
than going through nested method calls. It also makes the
TourStepAutomatic and TourAutomatic classes simpler and more
understandable.
(And this allows you to only have methods of less than 60 lines)

closes odoo/odoo#185542

X-original-commit: e85ee1d036fdd7398bb5909081111a5b7124cea0
Signed-off-by: Julien Mougenot (jum) <jum@odoo.com>
Signed-off-by: Pierre Pulinckx (pipu) <pipu@odoo.com>