[REF] web_tour: remove tour automatic callback from tour_service
In this commit, we move the tour_automatic callback in the tour_service
to the tour_automatic class. This allows for clearer handling of tour
error or success messages.
We take advantage of this commit to remove a small error
(clearTimeout(this._timeout)) in the tour_step_automatic where
this._timeout no longer exists.

closes odoo/odoo#186321

X-original-commit: 38c9dd1eb943fb8f97c97a3f45859eef4d77db3d
Signed-off-by: Bastien Fafchamps (bafa) <bafa@odoo.com>
Signed-off-by: Pierre Pulinckx (pipu) <pipu@odoo.com>