[REF] web_tour: macro.js
Before this commit, a macro can "move forward" either because the
previous step is completed or because of a DOM mutation. However,
one action or the other does not take precedence over the other.
Both can happen at the same time and cause the tour to "continue".
In other words, depending on how mutations arrive in the DOM, a tour
can behave differently from one time to the next.
After this commit, we force macro.js to have only one possible path.

taskId~4154597

closes odoo/odoo#186993

X-original-commit: 609f7b0d8e24cd8c99527b17f1dbc07bfa2a9ce6
Signed-off-by: Julien Mougenot (jum) <jum@odoo.com>
Signed-off-by: Pierre Pulinckx (pipu) <pipu@odoo.com>