[IMP] web_tour: add check for undeterminisms
In this commit, we add a new feature in tour engine that allows to check
if there are potential indeterminisms related to the triggers.
To do this, we will "simply" check if the targeted element (trigger) is
not altered within a delay.
If this is the case, it is then possible that we do not wait long
enough before targeting the element in the tour, or that the trigger
is not precise enough.

closes odoo/odoo#188533

X-original-commit: 421067ebaf926cb03716f20765f0bb057d5102a3
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>