[REF] web_tour: decrease default checkDelay value to 200ms
In macro.js, we check at each mutation of the DOM if the trigger is
present as soon as there has been no more mutation in the {checkDelay}
milliseconds. By default, this {checkDelay} is at 300ms in the
tourAutomatic class. In this commit, this default value is descreased
to 200ms for performance reasons.

closes odoo/odoo#192698

X-original-commit: 647ca0b6ef949ca61d7a031db51742ad0f5a8adc
Related: odoo/enterprise#76627
Signed-off-by: Bastien Pierre (ipb) <ipb@odoo.com>
Signed-off-by: Pierre Pulinckx (pipu) <pipu@odoo.com>