[FIX] web: fix unit test error_checker
Before this commit, the error_checker asserting that a JS unit test run
failed performed a strict equality with an outdated string on error
logs.

Now, error checker only checks for the [HOOT] prefix to determine
whether an error should end the test run, as to not rely on an arbitrary
error message.

closes odoo/odoo#196126

X-original-commit: 3049c1ea549290e934aa3ba24436af8b8d31dc05
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>
Signed-off-by: Julien Mougenot (jum) <jum@odoo.com>