[REF] web_tour: remove pointer in tour_automatic
In this commit, we remove the pointer from the tour_automatic.
Indeed, it has no useful use in this part of the code.
There is no point in having the pointer in the tour_automatic.
If we want to see the targeted element (trigger), just open the console.
Each element found is logged in the console in
debug mode (tour_automatic.js:65)
As the pointer is never used and has no use in the tour automatic,
we remove it.

closes odoo/odoo#197692

X-original-commit: 522107a7c9aa6f3bd3b14c0f46bb9d8c04283ff3
Signed-off-by: Bastien Fafchamps (bafa) <bafa@odoo.com>
Signed-off-by: Pierre Pulinckx (pipu) <pipu@odoo.com>